stage5-ml-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.2
numpy==1.26.4
pydantic==2.9.2
starlette-exporter==0.23.0
psycopg2-binary==2.9.9
python-json-logger==2.0.7
